+++ src/App.vue
@@ -2,6 +2,7 @@
 import { computed } from 'vue'
 import { RouterView, useRoute } from 'vue-router'
 import AppLayout from '@/components/layout/AppLayout.vue'
+import ToastContainer from '@/components/ToastContainer.vue'
 
 const route = useRoute()
 const isBlank = computed(() => (route.meta as any).layout === 'blank')
@@ -12,4 +13,5 @@ const isBlank = computed(() => (route.meta as any).layout === 'blank')
     <RouterView />
   </AppLayout>
   <RouterView v-else />
+  <ToastContainer />
 </template>